:recycle: Move transaction bytesizes into transaction constants
